requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pandas==2.2.2
pandera==0.18.3
pydantic==1.10.14
PyYAML==6.0.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.2
